pyproject.toml
[tool.poetry]
name = "openaiProject"
version = "0.1.0"
description = ""
authors = ["PostboxRetinal <78338192+PostboxRetinal@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
openai = "^1.42.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
